+++ include/datatype.hpp
@@ -1,9 +1,9 @@
 #ifndef PGFEM3D_DATATYPE
 #define PGFEM3D_DATATYPE
 
-using Ai_t = int; 
-using HYPRE_t = int;
-//using Ai_t = long;
-//using HYPRE_t = long long;
+//using Ai_t = int; 
+//using HYPRE_t = int;
+using Ai_t = long;
+using HYPRE_t = long long;
 
 #endif
